requirements.txt
ace_tools==0.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.58.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
virtualenv==20.31.2
Werkzeug==3.1.3
whitenoise==6.9.0
